requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Binance API
python-binance==1.0.19

# HTTP client
httpx==0.25.2

# Data analysis
pandas==2.1.4
numpy==1.26.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Environment variables
python-dotenv==1.0.0

# Async support
aiofiles==23.2.1

# Testing
pytest>=7.0.0
pytest-asyncio>=0.22.0
